platform/backend/pyproject.toml
[project]
name = "fintech-platform"
version = "0.1.0"
description = "AI-Native Fintech Infrastructure Platform — deterministic core + AI intelligence layer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "uvloop>=0.21.0; sys_platform != 'win32'",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "ulid-py>=1.1.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
ai = [
    "google-genai>=1.14.0",
    "instructor>=1.7.0",
    "tavily-python>=0.5.0",
    "firecrawl-py>=1.15.0",
]
ocr = [
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "faker>=33.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "packages"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
